Rename ReactSelect to SelectWithCreateOption
